requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
asgiref==3.7.2
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
av==11.0.0
azure-cognitiveservices-speech==1.35.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
contourpy==1.2.0
ctranslate2==4.0.0
cycler==0.12.1
dataclasses-json==0.6.4
decorator==4.4.2
distro==1.9.0
Django==5.0.2
docopt==0.6.2
einops==0.7.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
faiss-cpu==1.7.4
faster-whisper==1.0.0
flatbuffers==23.5.26
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.20.3
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.7
imageio==2.34.0
imageio-ffmpeg==0.4.9
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
julius==0.2.7
kiwisolver==1.4.5
langchain==0.1.9
langchain-community==0.0.22
langchain-core==0.1.26
langchain-openai==0.0.7
langsmith==0.1.5
lazy_loader==0.3
librosa==0.10.1
lightning==2.2.0.post0
lightning-utilities==0.10.1
llvmlite==0.42.0
Mako==1.3.2
markdown-it-py==3.0.0
marshmallow==3.20.2
matplotlib==3.8.3
mdurl==0.1.2
mkl-service==2.4.0
moviepy==1.0.3
msgpack==1.0.7
multidict==6.0.5
mypy-extensions==1.0.0
nltk==3.8.1
numba==0.59.0
omegaconf==2.3.0
onnxruntime==1.17.0
openai==1.12.0
openpyxl==3.1.2
optuna==3.5.0
packaging==23.2
pandas==2.2.0
pillow==10.3.0
platformdirs==4.2.0
pooch==1.8.1
primePy==1.3
proglog==0.1.10
protobuf==4.25.3
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytorch-lightning==2.2.0.post0
pytorch-metric-learning==2.4.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
semantic-version==2.10.0
semver==3.0.2
sentencepiece==0.2.0
setuptools-rust==1.8.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
speechbrain==0.5.16
SQLAlchemy==2.0.27
sqlparse==0.4.4
tabulate==0.9.0
tenacity==8.2.3
tensorboardX==2.6.2.2
threadpoolctl==3.3.0
tiktoken==0.6.0
tokenizers==0.15.2
tomli==2.0.1
torch==2.0.0
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.0.0
torchmetrics==1.3.1
torchvision==0.15.0
tqdm==4.66.2
transformers==4.38.1
typer==0.9.0
typing-inspect==0.9.0
tzdata==2023.4
urllib3==2.2.1
uuid==1.30
waitress==3.0.0
whisperx @ git+https://github.com/m-bain/whisperx.git@f2da2f858e99e4211fe4f64b5f2938b007827e17
yarl==1.9.4
python-dotenv
opneai